bot/requirements.txt
python-telegram-bot
docker
python-dotenv
